+++ tests/testthat/test-create_beast2_continue_cmd_from_options.R
@@ -34,3 +34,59 @@ test_that("Run with spaces in the ouput state filename, for Windows", {
   expect_false("file with spaces" %in% cmd)
   expect_true(shQuote("file with spaces") %in% cmd)
 })
+
+test_that("Run with spaces in the launcher.jar path, for Windows", {
+
+  if (!is_beast2_installed()) return()
+
+  beast2_path <- file.path(
+    tempfile(),
+    "path with spaces",
+    "launcher.jar"
+  )
+  expect_true(is_jar_path(beast2_path))
+  dir.create(dirname(beast2_path), showWarnings = FALSE, recursive = TRUE)
+  file.copy(
+    from = get_default_beast2_path(),
+    to = beast2_path
+  )
+  expect_true(file.exists(beast2_path))
+
+
+
+  cmd <- create_beast2_continue_cmd_from_options(
+    beast2_options = create_beast2_options(
+      beast2_path = beast2_path
+    )
+  )
+  expect_false(beast2_path %in% cmd)
+  expect_true(shQuote(beast2_path) %in% cmd)
+})
+
+test_that("Run with spaces in the beast bin path, for Windows", {
+
+  if (!is_beast2_installed()) return()
+
+  beast2_path <- file.path(
+    tempfile(),
+    "path with spaces",
+    "beast"
+  )
+  expect_true(is_bin_path(beast2_path))
+  dir.create(dirname(beast2_path), showWarnings = FALSE, recursive = TRUE)
+  file.copy(
+    from = get_default_beast2_bin_path(),
+    to = beast2_path
+  )
+  expect_true(file.exists(beast2_path))
+
+
+
+  cmd <- create_beast2_continue_cmd_from_options(
+    beast2_options = create_beast2_options(
+      beast2_path = beast2_path
+    )
+  )
+  expect_false(beast2_path %in% cmd)
+  expect_true(shQuote(beast2_path) %in% cmd)
+})