+++ tests/testthat/test-is_jar_path.R
@@ -2,6 +2,9 @@ context("is_jar_path")
 
 test_that("use", {
   expect_true(is_jar_path(get_default_beast2_jar_path()))
+  expect_true(is_jar_path("launcher.jar"))
+  expect_true(is_jar_path("path with spaces/launcher.jar"))
+  expect_true(is_jar_path("/path with spaces/more spaces/launcher.jar"))
   expect_false(is_jar_path(get_default_beast2_bin_path()))
   expect_false(is_jar_path("nonsense"))
   expect_false(is_jar_path(""))